Inline createProvisioningConfiguration
Additionally, always set provisioning timeout to 0 (disabled).
Provisioning with a static config always succeeds instantly, so
explicitly disabling the timeout is a noop.

Some of this functionality could / should probably be moved to
ProvisioningConfiguration directly (though it is slightly weird because
IpConfiguration is a bit of a misnomer as it also contains proxy
settings).

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: Ia211f9f1c54e1f1fe10142c9489e142818faef71