Start tracking local network requests
This change starts tracking requests for a local USB network. If a
request is added/removed, it will update the RequestTracker accordingly.
Note that local network requests take precedence. For as long as there
is a request for the local network, the provider will not bring up a
global network.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: If2323d7a2d700d1642ad4ba4ee166af0b5d8f905