Only validate TrackingReason when an interface is newly added
Only interfaces which are newly added should validate the
TrackingReason, since the TrackingReason may depend on the state of the
interface (the interface must still be present to lookup the driver via
ethtool API) or whether test interfaces are included.

This prevents ending up with orphaned interfaces because the RTM_DELLINK
event was not properly processed.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: Ie28865601b9b1e47458e816c965b03f1f0a4e40b