Update test to wait for LOST callback when disconnecting WiFi
testReportNetworkInterfaceForTransports is flaky, and one failure reason
is that the device fails to reconnect to Wi-Fi after a disconnect.
The current failure message is misleading because ensureWiFiConnected()
receives callbacks from the previous Wi-Fi network.
This CL replaces disableWifi with ensureWifiDisconnected, which
explicitly waits for the LOST callback.
While this isn't a direct fix for the test failure, it will be helpful
for investigating the issue.

Bug: 416687630
Test: BatteryStatsManagerTest
Flag: EXEMPT test only change
Change-Id: I27cdf87c465b8caf6d2b1cd41a133ca4b9d28377