Do not send InterfaceStateChanged callbacks for NCM-only interfaces
This change prevents triggering onInterfaceStateChanged callbacks for
interfaces which are not tracked in the regex (NCM-only).

To do so, the tracking reason is passed to EthernetListener which makes
the decision based on its mFlags whether to send the callback.
Unfortunately, the updateConfiguration APIs have also been triggering
onInterfaceStateChanged callbacks even on interfaces which are not being
tracked. To maintain this behavior, TrackingReason.REGEX is passed in
those instances. This is a bit of hack and will prevent us from passing
the InterfaceState object in the future (because obviously, untracked
interfaces do not have an associated InterfaceState object). Therefore,
we should consider removing this behavior, which feels inconsistent at
best (it is only triggered for IpConfiguration updates; not
NetworkCapabilities updates).

Test: TH
Flag: EXEMPT mainline
Change-Id: I4d7b386bf97cdc9207e2ee5d3595594de51f2784