Ignore null or empty interface names in EthernetTracker.
If EthernetTracker ever gets a null or empty interface name from a
netlink message, it will crash the system. This likely cannot happen,
but because the code runs all the time on all interfaces that have a
MAC address, it seems safer to have a check for it.

Test: None
Flag: EXEMPT bugfix
Change-Id: Iff93de9a42e5d8aa51e0183b45608c808ef31c38