BpfNetworkStats: optimize bpfGetIfaceStatsInternal() NULL/wildcard iface case
Note: this is not quite a no-op, as we no longer do the ifname
lookup, which could (but shouldn't) fail and cause us to ignore the stats
(for that interface).

But since it shouldn't ever fail, the value lookup should basically always
happen.  Switching to forAll() will allow bulk api optimization...

Test: TreeHugger, atest libbpf_android_test libnetworkstats_test netd_updatable_unit_test
Flag: EXEMPT mainline
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Id30fa1d2b4f30c5b526b5eba80e401d1a201d95d